PanasonicDecompressorV5: rewrite ctor without integer overflow UB

+++ src/librawspeed/decompressors/PanasonicDecompressorV5.h
@@ -62,7 +62,7 @@ private:
 
   const uint32 bps;
 
-  uint32 numBlocks;
+  size_t numBlocks;
 
   struct Block {
     ByteStream bs;
